# Speed-up contains by using `memchr` on every iteration